type: feature
switching from the reference implementation of the DDPG pendulum from Udacity (since it wasn't even converging locally for me on the original environment) to starting with the DDPG bipedal example. Already learned a bunch, had a day or so of troubleshooting hyperparams like learning rate and weight decay as well as reading through issues other students were having with lower rewards consistently for over 500 episodes like I was. None of them seemed to be the exact problem I was having. Debug printing my actions and noise level led to a realization. The initialization of the actor model was keeping all actions close to +1, so the arm was just swinging in a circle and never collecting reward because the ball is rarely in that location. I though tit was exploration noise related, but even with more noise and adding noise decay to move toward optimality it wasn't changing things. Having [1,1,1,1] for the clipped action prevented any exploration or learning. So, I changed the random seed (was 10 for the bipedal environment) down to 5 for the weights and  that brought things down to a more realistic motion for the torque values generated, allowed more exploration. Now will run a training run with these values and the noise decay slightly smaller and see what I get.