type feature
in an attempt to boost learning, I updated the type of random method used in the OUNoise in accordance with what I'd heard others have tried, though mentors said it shouldn't make a difference in the noise implementation. I also changed the network architecture and that seems to have resulted in more initial reward, but learning doesn't continue to improve.