type feature
did a lot of hyperparameter tuning around noise and learning rates and settled on these values for now, hoping this will lead to training a better policy. Doing a run now to test. Another, very important change, I changed the for loop for timesteps in an episode into a while true loop so it is only terminated when reaching the done state, this increases my episode length by 700 time steps which provides a lot more time for reward.